Code for Mars Rover (Mindstorms Ev4)
This code utilizes a Dualshock 4 Playstation 4 controller to control the Rover. Can be repeated with the  Xbox One controller and with a phone with minor tweaks. (Things to note: the program takes a while to load into the hub. Make sure the hub is connected via Bluetooth only and is placed in Streaming mode. you may find controlling it difficult because the Rover does not contain steering motors and uses tank coordination of the motors to turn. Anti-Collision protocols not included, be sure to supervise the robot while controlling. The front wheels move faster than the rear ones, due to gears. Front motors have been set at 33% output, rear ones have been set at max (100%) Fine-tune at your liking. Finally, the D-pad up button is used to make a high pitched beep, and the Cross and Triangle buttons can switch the speed between slow and fast motor outputs respectively)