Block Coded and Python Version of Fall Detection Software (6-26-2021)
This software is coded using both the Scratch Block feature and python in the Lego Mindstorms Ev4 software. The Scratch version of the code does pack more features than the Python version. However, I highly encourage you to use the Python version for smoother functionality, de-bugging, and to improve the code. DISCLAIMER: Must be used on an elevated surface over 12 inches from the surrounding surface. Best used on Rectangular surface. It will not work on pure black surfaces (dark colors are fine as long as they emit over 10% reflected light). In certain cases, the robot will fail to detect a fall using the color sensor, and a wheel might roll off the surface. This will shut down the motors and will beep until the program has been stopped. Also, if the robot detects there is no room for it to back up, it will beep until the program has been restarted. At startup, the Scratch version will show an animation, but the Python version will not. That's all I have to say for now. Have fun and Happy Coding!