Code that detects a fall and stops it at all costs.
The Robot may get in a sticky situation where the robot is on a diagonal trajectory over the edge, but the color sensor is unable to detect the fall. The robot will sense a fall using the built-in gyro and will stop. But it will not be able to get out of the position and beep for rescue. V1 is the super old and laggy version. Please do not look in there, it is embarrassing. V2 is the updated version coded more efficiently. V3 is designed for one sole purpose only, to conserve memory and reduce lag. The program ditches the Ultra-Sonic sensor and minimizes spacing, comments, and imports to conserve memory